+++ src/main/java/org/realityforge/replicant/client/EntityRepositoryDebugger.java
@@ -1,13 +1,47 @@
 package org.realityforge.replicant.client;
 
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.Collections;
+import java.util.Map;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
 public class EntityRepositoryDebugger
 {
   protected static final Logger LOG = Logger.getLogger( EntityRepositoryDebugger.class.getName() );
+
+  public void outputSubscription( final EntityRepository repository,
+                                  final EntitySubscriptionManager subscriptionManager,
+                                  final Enum graph,
+                                  final Class<?>[] types )
+  {
+    final ChannelSubscriptionEntry subscription = subscriptionManager.findSubscription( graph );
+    if ( null != subscription )
+    {
+      for ( final Class<?> type : types )
+      {
+        outputSubscriptionType( repository, subscription.getEntities(), type );
+      }
+    }
+  }
+
+  public void outputSubscriptionType( final EntityRepository repository,
+                                      final Map<Class<?>, Map<Object, EntitySubscriptionEntry>> entities,
+                                      final Class<?> type )
+  {
+    final Map<Object, EntitySubscriptionEntry> subscriptions = entities.get( type );
+    if ( null != subscriptions )
+    {
+      final Collection<EntitySubscriptionEntry> entries = subscriptions.values();
+      outputTypeHeader( type, entries.size() );
+      for ( final EntitySubscriptionEntry entry : entries )
+      {
+        final Object entity = repository.findByID( entry.getType(), entry.getID() );
+        outputEntity( entry.getType(), entry.getID(), entity );
+      }
+    }
+  }
 
   public void outputRepository( final EntityRepository repository )
   {